Corrige 4 bugs critiques et ajoute la réinitialisation de mot de passe
- Logements invisibles: le filtre prix envoyait un double (60000.0) à une
  colonne integer, faisant échouer toutes les requêtes en 400
- Ordre des messages inversé: .order('date_envoi') sans ascending:true
  triait par défaut en descendant côté postgrest-dart
- Déconnexion qui nécessitait deux tentatives: signOut() sans gestion
  d'erreur bloquait la navigation en cas de souci réseau
- Se souvenir de moi ne conservait pas le mot de passe
- Ajout du flux complet mot de passe oublié / réinitialisation par email
- Redirection automatique fiabilisée après connexion/inscription Google

+++ lib/features/logement/screens/logement_screen.dart
@@ -48,7 +48,7 @@ class _LogementScreenState extends State<LogementScreen> {
         query = query.eq('type', _selectedType);
       }
 
-      query = query.lte('prix', _prixMax);
+      query = query.lte('prix', _prixMax.toInt());
 
       if (_searchController.text.isNotEmpty) {
         query = query.or(
